+++ src/lexer/tokens.ts
@@ -104,4 +104,10 @@ export const getKeywordType = (value: string): GomToken => {
   }
 };
 
-export const GOM_BUILT_IN_TYPES = new Set(["i8", "bool", "f32", "str", "void"]);
+export const GOM_BUILT_IN_TYPES = new Set([
+  "int",
+  "bool",
+  "float",
+  "str",
+  "void",
+]);
